Added waves, updated GUI

+++ Assets/Scripts/Building/BuildingStats.cs
@@ -11,7 +11,7 @@ public class BuildingStats:MonoBehaviour {
 	public int[] stoneCostPerLevel;
 	public int[] woodCostPerLevel;
 
-	public int[] rangePerLevel;
-	public int[] damagePerLevel;
-	public int[] speedPerLevel;
+	public float[] rangePerLevel;
+	public float[] damagePerLevel;
+	public float[] speedPerLevel;
 }
